TunnelFactory web api (#88200)
Web API for tunnels
Part of #81388
