+++ src/vs/workbench/contrib/remote/common/tunnelFactory.ts
@@ -1,0 +1,39 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { ITunnelService, TunnelOptions, RemoteTunnel } from 'vs/platform/remote/common/tunnel';
+import { Disposable } from 'vs/base/common/lifecycle';
+import { IWorkbenchContribution } from 'vs/workbench/common/contributions';
+import { IWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
+
+export class TunnelFactoryContribution extends Disposable implements IWorkbenchContribution {
+	constructor(
+		@ITunnelService tunnelService: ITunnelService,
+		@IWorkbenchEnvironmentService workbenchEnvironmentService: IWorkbenchEnvironmentService,
+	) {
+		super();
+		if (workbenchEnvironmentService.options && workbenchEnvironmentService.options.tunnelFactory) {
+			this._register(tunnelService.setTunnelProvider({
+				forwardPort: (tunnelOptions: TunnelOptions): Promise<RemoteTunnel> | undefined => {
+					const tunnelPromise = workbenchEnvironmentService.options!.tunnelFactory!(tunnelOptions);
+					if (!tunnelPromise) {
+						return undefined;
+					}
+					return new Promise(resolve => {
+						tunnelPromise.then(tunnel => {
+							const remoteTunnel: RemoteTunnel = {
+								tunnelRemotePort: tunnel.remoteAddress.port,
+								tunnelRemoteHost: tunnel.remoteAddress.host,
+								localAddress: tunnel.localAddress,
+								dispose: tunnel.dispose
+							};
+							resolve(remoteTunnel);
+						});
+					});
+				}
+			}));
+		}
+	}
+}
