+++ src/vs/workbench/workbench.web.api.ts
@@ -34,6 +34,26 @@ interface IExternalUriResolver {
 	(uri: URI): Promise<URI>;
 }
 
+interface TunnelOptions {
+	remoteAddress: { port: number, host: string };
+	// The desired local port. If this port can't be used, then another will be chosen.
+	localAddressPort?: number;
+	label?: string;
+}
+
+interface Tunnel {
+	remoteAddress: { port: number, host: string };
+	//The complete local address(ex. localhost:1234)
+	localAddress: string;
+	// Implementers of Tunnel should fire onDidDispose when dispose is called.
+	onDidDispose: Event<void>;
+	dispose(): void;
+}
+
+interface ITunnelFactory {
+	(tunnelOptions: TunnelOptions): Thenable<Tunnel> | undefined;
+}
+
 interface IWorkbenchConstructionOptions {
 
 	/**
@@ -103,6 +123,11 @@ interface IWorkbenchConstructionOptions {
 	 * Resolves an external uri before it is opened.
 	 */
 	readonly resolveExternalUri?: IExternalUriResolver;
+
+	/**
+	 * Support for creating tunnels.
+	 */
+	readonly tunnelFactory?: ITunnelFactory;
 
 	/**
 	 * Current logging level. Default is `LogLevel.Info`.
